add rpia_stream script for streaming
This scales the video down to 720p because these streams do not seem to work
with 1080p. The command line argument may be a stream key or a full RTMP
URL (starting with rtmp:) that will be passed directly to ffmpeg. If only
the stream key is provided, rtmp://va.pscp.tv:80/x/ is assumed as the url
prefix.